Move undo/redo hook to hooks directory
